backend/requirements.txt
fastapi
uvicorn
langgraph
psycopg2-binary
sentence-transformers
langchain-community
langchain-ollama
requests